- clean up examples with type casts where needed. - clean up main docs to match landing page.

+++ py/examples/04_moderate_content.py
@@ -67,7 +67,7 @@ store_moderation_result = Activity(
         "query": """
             INSERT INTO moderation_log
             (content_id, decision, cost, tokens, moderated_at)
-            VALUES ($1, $2, $3, $4, NOW())
+            VALUES ($1, $2, $3::numeric, $4, NOW())
         """,
         "params": [
             str(content_id),
